requirements.txt
# PhotoFlow - Core requirements
PyQt6>=6.6.0
Pillow>=10.0.0
cryptography>=41.0.0

# AI analysis (local, via Ollama - run: ollama pull llava)
ollama>=0.1.0

# Image processing
numpy>=1.24.0
opencv-python>=4.8.0

# Photo organization
imagehash>=4.3.1          # Perceptual hashing for duplicate detection

# Social media APIs
requests>=2.31.0
requests-oauthlib>=1.3.1  # OAuth2 for Instagram Graph, Twitter, Facebook, Pinterest

# Optional: Twitter/X client
# tweepy>=4.14.0

# Optional: folder auto-import watcher
# watchdog>=4.0.0

# Removed (face recognition - Nova-specific, not needed for general use):
# deepface>=0.0.79
# opencv-contrib-python>=4.8.0
# tensorflow / tf-keras
